Renamed package, added logs, changed expected type of activities.
