+++ src/main/java/school/faang/bjs268102/Main.java
@@ -1,5 +1,4 @@
-package school.faang.BJS2_68102;
-
+package school.faang.bjs268102;
 
 public class Main {
     public static void main(String[] args) {
